Fixed start distance in Example shader to begin at AABB.
In the Example shader the ray origin is camera position and the start distance was zero. This incurred additional steps to be taken as incrememnted distance was based on distance from AABB to SDF. Now distance starts as distance from camera to AABB. This can greatly reduce number of steps taken per ray.